// src/generated-api/schemas/hackathonResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { HackathonResponseType } from './hackathonResponseType';
import type { HackathonResponseHackathonState } from './hackathonResponseHackathonState';

export interface HackathonResponse {
  id?: number;
  name?: string;
  description?: string;
  type?: HackathonResponseType;
  hackathonState?: HackathonResponseHackathonState;
  startedAt?: string;
  endedAt?: string;
}